+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 * Change heading of modal generate accession and call number. Refs UIIN-3274.
 * ViewSource - add tenantId to useAuditSettings. Refs UIIN-3266.
 * Enhancement help text on Settings > Inventory > Number generator options. Refs UIIN-3217.
+* Add default value for number of versions counter on "Version history" pane. Fixes UIIN-3283.
 
 ## [12.0.12](https://github.com/folio-org/ui-inventory/tree/v12.0.12) (2025-01-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.11...v12.0.12)